# contracts/openapi.yaml
{
  "openapi": "3.1.1",
  "info": {
    "title": "Voice Platform API",
    "version": "1.0.0-draft",
    "description": "Versioned foundation contract. Feature endpoints are added only with matching ACL and contract tests."
  },
  "paths": {
...
    "/api/v1/direct-messages/{directMessageID}/search": {
      "get": {
        "operationId": "searchDirectMessageHistory",
        "description": "Authenticated DM participant only. Searches non-deleted message text in the caller's selected canonical pair with PostgreSQL simple full-text search; administrator role grants no exception.",
        "parameters": [
          { "name": "directMessageID", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "query", "in": "query", "required": true, "schema": { "type": "string", "minLength": 1, "maxLength": 256 } },
          { "name": "before", "in": "query", "required": false, "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 } }
        ],
        "responses": {
          "200": { "description": "Newest-first matching message page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DirectMessageSearchPage" } } } },
          "400": { "description": "Malformed query or page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "No active server-side session", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "DM is absent or caller is not a participant", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Internal error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/v1/voice/channels/{channelID}/leases": {
      "post": {
        "operationId": "acquireVoiceLease",
        "description": "Authenticated user only. Creates one logical lease for an active VOICE channel. A second tab or device receives 409 until the user explicitly submits transfer:true; this command does not itself claim that LiveKit is connected.",
        "parameters": [
          { "name": "channelID", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VoiceLeaseRequest" } } }
        },
        "responses": {
          "201": { "description": "Logical lease created; media credential and LiveKit confirmation are separate steps", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VoiceLease" } } } },
          "400": { "description": "Malformed request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "No active server-side session", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "409": { "description": "Existing lease, closed admission, archived/non-voice channel or stale state", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VoiceLeaseConflict" } } } },
          "500": { "description": "Internal error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/v1/admin/accounts/{accountID}/voice-kick": {
      "post": {
        "operationId": "kickVoiceParticipant",
        "description": "Administrator-only. Revokes the target account's current logical voice lease, records audit metadata and creates a durable SFU revocation. The response counts logical leases only; completion and token replay remain POC-03 gates.",
        "parameters": [
          { "name": "accountID", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Logical leases revoked", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VoiceKick" } } } },
          "400": { "description": "Malformed account ID", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "No valid active session", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "403": { "description": "Administrator role required", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Internal error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DirectMessageSearchPage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["messages"],
        "properties": {
          "messages": { "type": "array", "items": { "$ref": "#/components/schemas/DirectMessageSearchResult" } },
          "next_cursor": { "type": "string", "format": "uuid" }
        }
      },
      "DirectMessageSearchResult": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "direct_message_id", "author_id", "body", "created_at", "revision"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "direct_message_id": { "type": "string", "format": "uuid" },
          "author_id": { "type": "string", "format": "uuid" },
          "body": { "type": "string", "minLength": 1, "maxLength": 8000 },
          "created_at": { "type": "string", "format": "date-time" },
          "edited_at": { "type": "string", "format": "date-time" },
          "revision": { "type": "integer", "minimum": 1 }
        }
      },
...
      "VoiceLeaseRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "transfer": { "type": "boolean", "default": false, "description": "Must be true to revoke this account's existing active lease." }
        }
      },
      "VoiceLease": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "channel_id", "transferred"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "channel_id": { "type": "string", "format": "uuid" },
          "transferred": { "type": "boolean" }
        }
      },
      "VoiceLeaseConflict": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorDetail" },
          "active_channel_id": { "type": "string", "format": "uuid" }
        }
      },
      "VoiceKick": {
        "type": "object",
        "additionalProperties": false,
        "required": ["revoked_leases"],
        "properties": {
          "revoked_leases": { "type": "integer", "minimum": 0 }
        }
      },
...
      "ErrorDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message", "request_id"],
        "properties": {
          "code": {
            "type": "string",
            "enum": ["UNAUTHENTICATED", "ACCOUNT_BLOCKED", "FORBIDDEN", "NOT_FOUND", "CONFLICT", "VALIDATION_FAILED", "RATE_LIMITED", "MAINTENANCE", "MAINTENANCE_UNAVAILABLE", "ATTACHMENT_TOO_LARGE", "INSUFFICIENT_STORAGE", "ACTIVE_VOICE_LEASE", "VOICE_CHANNEL_UNAVAILABLE", "VOICE_LEASE_UNAVAILABLE", "INTERNAL"]
          },
          "message": { "type": "string" },
          "request_id": { "type": "string" }
        }
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorDetail" }
        }
      }
    }
  }
}